requirements.txt
llama-cpp-python
openai-whisper
pyttsx3
sounddevice
numpy
scipy
